// src/app/share/wy-ui/wy-slider/wy-slider-types.ts
import { Observable } from 'rxjs';

export interface WySliderStyle {
    width?: string | null;
    height?: string | null;
    left?: string | null;
    bottom?: string | null;
}

export interface SliderEventObserverConfig {
    start: string;
    move: string;
    end: string;
    filterHandle: (e: Event) => boolean;
    pluckKey: string[];
    startPlucked$?: Observable<number>;
    moveResolved$?: Observable<number>;
    end$?: Observable<Event>;
}

export type SliderValue = number | null;